+++ client/action-types.ts
@@ -1,7 +1,6 @@
 import { AuthActions } from './auth/actions'
 import { ChatActions } from './chat/actions'
 import { DialogActions } from './dialogs/actions'
-import { FileBrowserActions } from './file-browser/actions'
 import { GamesActions } from './games/actions'
 import { LadderActions } from './ladder/actions'
 import { MapsActions } from './maps/actions'
@@ -15,7 +14,6 @@ type AllActions =
   | AuthActions
   | ChatActions
   | DialogActions
-  | FileBrowserActions
   | GamesActions
   | LadderActions
   | MapsActions
