+++ client/reducers.ts
@@ -6,7 +6,6 @@ export { default as matchmakingTimes } from './admin/matchmaking-times-reducer'
 export { default as auth } from './auth/auth-reducer'
 export { default as chat } from './chat/chat-reducer'
 export { default as dialog } from './dialogs/dialog-reducer'
-export { default as fileBrowser } from './file-browser/file-browser-reducer'
 export { default as gameplayActivity } from './gameplay-activity/gameplay-activity-reducer'
 export { default as games } from './games/game-reducer'
 export { default as ladder } from './ladder/ladder-reducer'
